refactor: Created frontend branch for deploying to s3
Creates the branch frontend which is designed for deploying to s3. In this branch, a condensed moduleData.json is located in the dist directory, with the LandingPage.js file edited to direct to this file. This removes the need for using the express server to deploy to production. For deploying locally, the express server will still be required, and the LandingPage.js will need to be edited back. See the updated README.md for more details.